# [SYSTEMDS-3390] Improve performance of countDistinctApprox()